UPDATE UNIT TESTS, FINAL CHECKS DONE
