+++ src/index.js
@@ -4,7 +4,7 @@ import { getShowsData } from './modules/APImethods.js';
 import {
   getButtonsList, getLikesCounters, populateLikes, renderShows, setClickListeners,
 } from './modules/UImethods.js';
-import detailsButton from './modules/detailsButton.js';
+import commentsButton from './modules/commentsButton.js';
 
 const showsIds = ['tt5491994', 'tt0903747', 'tt0795176', 'tt0185906', 'tt7366338', 'tt0306414'];
 
@@ -14,4 +14,4 @@ renderShows(showsData);
 setClickListeners(getButtonsList());
 populateLikes(getLikesCounters());
 
-detailsButton();
+commentsButton();
