+++ src/modules/commentsButton.js
@@ -1,6 +1,8 @@
-// import getShowsData from './APImethods.js';
+import comments from './comments.js';
 
-const popup = () => {
+const baseUrl = 'https://api.tvmaze.com/lookup/shows?imdb=';
+
+const popup = (Obj) => {
   const container = document.createElement('div');
   container.className = 'popup-container';
 
@@ -12,14 +14,38 @@ const popup = () => {
   });
   container.appendChild(closeIcon);
 
-  return container;
+  const fig = document.createElement('FIGURE');
+  fig.className = 'show-image';
+  container.appendChild(fig);
+
+  const img = document.createElement('img');
+  img.setAttribute('src', Obj.image.medium);
+  img.setAttribute('width', 'auto');
+  img.setAttribute('height', 'auto');
+
+  fig.appendChild(img);
+
+  const showname = document.createElement('div');
+  showname.className = 'show-name';
+  showname.innerText = Obj.name;
+  container.appendChild(showname);
+
+  document.body.appendChild(container);
+
+  comments(Obj);
+};
+
+const retriveData = async (imdbId) => {
+  const response = await fetch(baseUrl + imdbId);
+  const Obj = await response.json();
+  popup(Obj);
 };
 
 const commentsButton = () => {
-  const btns = document.querySelectorAll('#commentsButton');
+  const btns = document.querySelectorAll('.commentsButton');
   btns.forEach((btn) => {
     btn.addEventListener('click', () => {
-      document.body.appendChild(popup());
+      retriveData(btn.id);
     });
   });
 };
