+++ src/modules/commentsCounter.js
@@ -1,0 +1,3 @@
+const commentsCounter = (ArrayObj) => ArrayObj.length;
+
+export default commentsCounter;
